requirements.txt
pillow
requests
customtkinter
google-genai